refactor: nomenclature pipeline segments structure
